requirement2.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
beautifulsoup4==4.10.0
buttons==0.1.9
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
colour==0.1.5
discord-py-slash-command==3.0.3
discord.py==1.7.3
distlib==0.3.4
filelock==3.4.2
google==3.0.0
idna==3.3
lyricsgenius==3.0.1
MouseInfo==0.1.3
multidict==6.0.2
numpy==1.22.2
opencv-python==4.5.5.62
packaging==21.3
Panda3D==1.10.11
panda3d-gltf==0.13
panda3d-simplepbr==0.9
phonenumbers==8.12.45
Pillow==9.0.1
pipenv==2022.1.8
platformdirs==2.4.1
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.6
pyparsing==3.0.7
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.28
pytesseract==0.3.9
python-dotenv==0.19.2
pytweening==1.0.4
pywin32==303
requests==2.27.1
screeninfo==0.8
six==1.16.0
soupsieve==2.3.1
typing_extensions==4.1.1
urllib3==1.26.8
ursina==4.1.1
virtualenv==20.13.0
virtualenv-clone==0.5.7
WMI==1.5.1
yarl==1.7.2